feat: add game room join event, delete game room scheduler #66
