+++ src/main/java/com/kospot/application/multiplayer/gameroom/usecase/LeaveGameRoomUseCase.java
@@ -1,4 +1,4 @@
-package com.kospot.application.multiplayer.gameroom;
+package com.kospot.application.multiplayer.gameroom.usecase;
 
 import com.kospot.domain.member.entity.Member;
 import com.kospot.domain.multigame.gameRoom.adaptor.GameRoomAdaptor;
@@ -6,7 +6,6 @@ import com.kospot.domain.multigame.gameRoom.entity.GameRoom;
 import com.kospot.domain.multigame.gameRoom.service.GameRoomService;
 import com.kospot.infrastructure.annotation.usecase.UseCase;
 import com.kospot.infrastructure.websocket.domain.gameroom.service.GameRoomPlayerService;
-import com.kospot.presentation.multigame.gameroom.dto.request.GameRoomRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,18 +14,20 @@ import org.springframework.transaction.annotation.Transactional;
 @UseCase
 @RequiredArgsConstructor
 @Transactional
-public class JoinGameRoomUseCase {
+public class LeaveGameRoomUseCase {
 
     private final GameRoomAdaptor gameRoomAdaptor;
     private final GameRoomService gameRoomService;
     private final GameRoomPlayerService gameRoomPlayerService;
 
-    public void execute(Member player, Long gameRoomId, GameRoomRequest.Join request) {
+    public void execute(Member player, Long gameRoomId) {
         GameRoom gameRoom = gameRoomAdaptor.queryById(gameRoomId);
-        // 데이터베이스 레벨에서 입장 처리
-        gameRoomService.joinGameRoom(player, gameRoom, request);
         
-        // 주의: 실시간 알림은 WebSocket 구독 시점에서 GameRoomSessionManager.addSubscription()을 통해 처리됩니다.
+        // 데이터베이스 레벨에서 퇴장 처리
+        gameRoomService.leaveGameRoom(player, gameRoom);
+        
+        // WebSocket 레벨에서 실시간 퇴장 처리 (Redis + 실시간 알림)
+        gameRoomPlayerService.removePlayerFromRoom(gameRoomId, player.getId());
     }
 
 }
