+++ src/main/java/com/kospot/domain/multigame/gameRoom/event/GameRoomJoinEvent.java
@@ -1,0 +1,15 @@
+package com.kospot.domain.multigame.gameRoom.event;
+
+import com.kospot.domain.member.entity.Member;
+import com.kospot.domain.multigame.gameRoom.entity.GameRoom;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class GameRoomJoinEvent {
+
+    private final GameRoom gameRoom;
+    private final Member member;
+
+}
